+++ packages/dialtone-vue2/components/radio_group/radio_group.vue
@@ -3,7 +3,7 @@ import { DtInputGroup } from '../input_group';
 
 /**
  * Radio Groups are control elements that allow the user to make a single selection from a list of options.
- * @see https://dialpad.design/components/radio_group.html
+ * @see https://dialtone.dialpad.com/components/radio_group.html
  */
 export default {
   name: 'DtRadioGroup',