+++ packages/dialtone-vue2/components/rich_text_editor/rich_text_editor.stories.js
@@ -7,6 +7,8 @@ import {
   RICH_TEXT_EDITOR_AUTOFOCUS_TYPES,
 } from './rich_text_editor_constants';
 
+import mentionSuggestion from './mention_suggestion';
+
 // Default Prop Values
 export const argsData = {
   value: `I’m not a standalone component, please use Message Input instead 🙏!`,
@@ -14,8 +16,8 @@ export const argsData = {
   inputAriaLabel: 'This is a descriptive label',
   outputFormat: 'text',
   autoFocus: false,
-  link: true,
   placeholder: 'Type here...',
+  link: true,
   onBlur: action('blur'),
   onInput: action('input'),
   onFocus: action('focus'),
@@ -99,10 +101,18 @@ export const Default = {
 };
 
 export const WithLinks = {
-  ...Default,
+  render: (argsData) => createRenderConfig(DtRichTextEditor, DtRichTextEditorDefaultTemplate, argsData),
   args: {
     link: true,
     value: 'The editor can autolink URLs: dialpad.com, https://www.dialpad.com/about-us/, ' +
     'IP addresses: 192.158.1.38, email addresses: noreply@dialpad.com and phone numbers: (778) 765-8813, +17787658813!',
+  },
+};
+
+export const WithMentionSuggestions = {
+  render: (argsData) => createRenderConfig(DtRichTextEditor, DtRichTextEditorDefaultTemplate, argsData),
+  args: {
+    value: 'The editor can also suggest mentions: @John Doe, @Jane Doe!',
+    mentionSuggestion,
   },
 };