+++ packages/dialtone-vue2/components/tabs/tab.vue
@@ -29,7 +29,7 @@ import { DtButton } from '../button';
 
 /**
  * Tabs allow users to navigation between grouped content in different views while within the same page context.
- * @see https://dialpad.design/components/tabs.html
+ * @see https://dialtone.dialpad.com/components/tabs.html
  */
 export default {
   name: 'DtTab',