+++ packages/dialtone-vue2/vite.config.js
@@ -15,6 +15,7 @@ export default defineConfig({
       entry: {
         'dialtone-vue': resolve(__dirname, './index.js'),
         emoji: resolve(__dirname, './emoji.js'),
+        message_input: resolve(__dirname, './message_input.js'),
         directives: resolve(__dirname, './directives.js'),
       },
     },